fix: moved `transformCode`
